always require internet to download installer

+++ system/ui/mici_setup.py
@@ -500,7 +500,7 @@ class Setup(Widget):
 
     self._network_setup_page = NetworkSetupPage(self._network_monitor, self._network_setup_continue_callback, self._pop_to_software_selection)
 
-    self._software_selection_page = SoftwareSelectionPage(self._use_openpilot, lambda: gui_app.push_widget(self._custom_software_warning_page))
+    self._software_selection_page = SoftwareSelectionPage(self._push_network_setup, lambda: gui_app.push_widget(self._custom_software_warning_page))
 
     self._download_failed_page = FailedPage(self._pop_to_software_selection, icon="icons_mici/setup/red_warning.png")
 
@@ -530,7 +530,7 @@ class Setup(Widget):
     gui_app.pop_widgets_to(self._software_selection_page, self._software_selection_page.reset)
 
   def _use_openpilot(self):
-    if os.path.isdir(INSTALL_PATH) and os.path.isfile(VALID_CACHE_PATH):
+    if os.path.isdir(INSTALL_PATH) and os.path.isfile(VALID_CACHE_PATH) and False:
       os.remove(VALID_CACHE_PATH)
       # with open(TMP_CONTINUE_PATH, "w") as f:
       #   f.write(CONTINUE)
@@ -538,7 +538,7 @@ class Setup(Widget):
       #  make sure installer fails (doesn't write continue.sh) if fetch/checkout fails
       #  make isntaller move atomic
       # run_cmd(["chmod", "+x", TMP_CONTINUE_PATH])
-      shutil.move(TMP_CONTINUE_PATH, CONTINUE_PATH)
+      # shutil.move(TMP_CONTINUE_PATH, CONTINUE_PATH)
       # shutil.copyfile(INSTALLER_SOURCE_PATH, INSTALLER_DESTINATION_PATH)
       shutil.copyfile(INSTALLER_SOURCE_PATH, TMP_INSTALLER_PATH)
       os.rename(TMP_INSTALLER_PATH, INSTALLER_DESTINATION_PATH)
@@ -618,13 +618,14 @@ class Setup(Widget):
         self._download_failed_reason = "No custom software found at this URL: " + self.download_url.replace("https://", "", 1)
         return
 
+      # TODO: this is unused?!
+      with open(INSTALLER_URL_PATH, "w") as f:
+        f.write(self.download_url)
+
       # AGNOS might try to execute the installer before this process exits.
       # Therefore, important to close the fd before renaming the installer.
       os.close(fd)
       os.rename(tmpfile, INSTALLER_DESTINATION_PATH)
-
-      with open(INSTALLER_URL_PATH, "w") as f:
-        f.write(self.download_url)
 
       # give time for installer UI to take over
       time.sleep(0.1)
